requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
Bootstrap-Flask==2.4.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dominate==2.9.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector==2.2.9
mysqlclient==2.2.6
packaging==24.2
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.1.0
soupsieve==2.6
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.3.0
visitor==0.1.3
Werkzeug==3.1.3
wheel==0.44.0
WTForms==3.2.1
WTForms-SQLAlchemy==0.4.2
